Fixed biome id error on chunk axis x=0, z=0
The root cause of the problem is unjustified, but the changes made here are to make sure that SimpleChunkManager destructed using the C++ class destructors rather than destroying objects in the function itself. This does fix the biome to generate its values properly (The internals biome generator works flawlessly without any issues). And of course, "If it works, it works".

+++ lib/objects/BiomeArray.cpp
@@ -6,23 +6,23 @@ BiomeArray::BiomeArray(const gsl::span<const uint8_t, 256> &values) {
 }
 
 uint8_t BiomeArray::get(const uint8_t x, const uint8_t z) const {
-  unsigned int offset;
+  uint_fast8_t offset;
   index(x, z, offset);
 
   return mValues[offset];
 }
 
 void BiomeArray::set(const uint8_t x, const uint8_t z, const uint8_t value) {
-  unsigned int offset;
+  uint_fast8_t offset;
   index(x, z, offset);
 
   mValues[offset] = value;
 }
 
-void BiomeArray::index(const uint8_t x, const uint8_t z, unsigned int &offset) {
+void BiomeArray::index(const uint8_t x, const uint8_t z, uint_fast8_t &offset) {
   offset = (z << 4) | x;
 }
 
-gsl::span<const uint8_t, BiomeArray::DATA_SIZE> BiomeArray::getRawData() const {
-  return gsl::span<const uint8_t>(mValues);
+gsl::span<const uint_fast8_t, BiomeArray::DATA_SIZE> BiomeArray::getRawData() const {
+  return gsl::span<const uint_fast8_t>(mValues);
 }
